Now when swimming, the body can drift forward a bit (but the joysticks must be released or the armTip will still drag the body behind).
Need to add in the factor of the angle between the two arms (if the angle between two arms are too small, it means they are all stretch towards one direction instead of spreading out to opposite directions like actually swimming), the smaller the angle, the less force it should provide to push the body to swim.

The first segment of the arms (the ones that are closest to the body center) tends to be difficult to move around its joint that connects to the center of the body (when the arm is suddenly pulling to the opposite direction, since the first segment is connected to the center and have to turn 180 degree, it's hard for it to turn since it cannot just flip its direction to the opposite direction).