Fixed a bug where sometimes the trigger status is updated before some functions leading to the problems that the arm cannot pick up items.
Temp switch render pipeline from light-weight to normal. (Transparent objects have no shadows in light-weight render pipeline)

Removed the default air glide force so that the wind force applied on the arm is calculated accurately. Need to base the force on the actual arm in wind length (now as long as the armTip is in the wind the entire arm will be calculated)

Keep improving mechanical arm collision detection.