Keep working on player soft body. Only minor gitters left when do a very sharp turn with the Compressed Air Cylinder. Consider clamp the maximum "joint damper, massScale" and "vertex rigidbody angular drag" and "Last  Arm Segment Connected Mass Scale" changing rate to get rid of the gittering (it's gittering because those parameters are changing too fast).
When have time, need to work to make the soft body not "too heavy", not it feels like the player has to drag their body really hard to move the body. Also, the soft body feels too "soft" now, need to make it less "soft" so it doesn't feel like a drop of liquid.